refactor: extract SceneSetupManager from IDE.java (#706)
Move scene-setup code generation and field reorganization logic into
a new package-private SceneSetupManager delegate class. This reduces
IDE.java from 539 to 432 lines (well under the 500-line target).

Extracted:
- UnacceptableFieldAccessCrawler inner class
- reorganizeTypeFieldsIfNecessary method
- reorganizeFieldsIfNecessary method
- generateCodeForSceneSetUp method

IDE.updateProject() now delegates to sceneSetupManager.
No public API changes. All 1357 module tests pass.
